// gig-rdo-vite/src/types/em-tempo-real-types.ts
export interface EmTempoRealResumo {
  total: number;
  atendimento: number;
  liberada: number;
  pendente: number;
  baixa_estoque: number;
  entrada_estoque: number;
  sem_movimentacao: number;
  reserva_estoque: number;
  valor_total: number;
}